requirements.txt
streamlit==0.82.0
plotly==4.14.3
numpy==1.19.2
pandas==1.1.3
Pillow==8.2.0
